Convert UndoIcon and RedoIcon to typescript.
